Changed Command.java class into an interface and abstracted all the various commands into their own for more OOP, as per suggested by @CheyanneSim
