+++ src/main/java/duke/command/ByeCommand.java
@@ -1,0 +1,8 @@
+package duke.command;
+
+public class ByeCommand implements Command {
+    //Custom byeCommand Method to exit Duke
+    public static String run() {
+        return ("BYE");
+    }
+}
